requirements.txt
# Production dependencies for Vercel Functions
fastapi==0.104.1
python-multipart==0.0.6
pillow==10.1.0
rembg==2.0.58
mangum==0.17.0
pydantic==2.5.0